CryptoDatasets

Documentation for CryptoDatasets.

Hello, World!

CryptoDatasets.Candle

CryptoDatasets._import_json!

CryptoDatasets._sanitize

CryptoDatasets.dataset

CryptoDatasets.Candle — Type

A candle containing a timestamp, OHLCV values, and an optional v2 for weird exchanges that return 2 volumes.

CryptoDatasets._import_json! — Method

importjson!(exchange, market, timeframe)

Import data from a previous project of mine that stored this info in JSON files.

CryptoDatasets._sanitize — Method

_sanitize()

Take a row from the JSON import and make it suitable for writing out as a CSV row.

Examples

julia> CryptoDatasets._sanitize([1636675140000, 1, 2, 0.5, 1, 1])
7-element Vector{Any}:
 1.63667514e12
 1.0
 2.0
 0.5
 1.0
 1.0
  missing

CryptoDatasets.dataset — Method

dataset(exchange, market; tf::Period, dates) => Vector{Any}

Return OHLC candles for the given exchange and market.